update performance - typecheck Anchors 50ms
